CI spec excludes updates.
* Fixed rake pristine task to whitelist Kernel#require fixtures.
* Updated CI excludes for Dir.glob and Dir.[].

+++ Rakefile
@@ -370,6 +370,7 @@ desc "Remove all stray compiled Ruby files"
 task :pristine do
   FileList['**/*.rbc'].each do |fn|
     next if /^runtime/.match(fn)
+    next if /require\/require_spec_.\.rbc/.match(fn)
     FileUtils.rm fn rescue nil
   end
 end
